* Makefile: adapted makefile to carry a DEBUG= option for compiling   flexcat without any debugging enabled at all. Also added the "-m32" gcc   option to the unix builds so that a 32bit version of flexcat will be   generated. We somehow still have problems on getting correct   catalog files generated when creating them with a 64bit flexcat. * flexcat.h, misc: introduced own portable datatype definitions like   uint32,int32 and so on. These should help to nail down existing problems   when building flexcat on 64bit system where e.g. long is defined as a   64bit value. Adapted all 'unsigned long' and similar 64bit usages to   use these new types instead.

+++ src/createcatsrc.c
@@ -384,7 +384,7 @@ void TerminateCatStringOutput ( void )
 
 /* This writes a source string. */
 
-void WriteString ( FILE * fpout, char *str, long Len, int lenbytes )
+void WriteString ( FILE * fpout, char *str, int32 Len, int lenbytes )
 {
     char            bytes[10];
     int             bytesread;
@@ -739,7 +739,7 @@ void CreateSourceFile ( char *SourceFile, char *TemplateFile, char *CDFile )
                                         char            _StrLen[20 + 1];
 
                                         sprintf ( _StrLen, "%020lx",
-                                                  ( unsigned long )cs->ID );
+                                                  ( uint32 )cs->ID );
                                         start = &_StrLen[20 - _len * 2];
                                         while ( _len > 0 )
 
@@ -760,8 +760,7 @@ void CreateSourceFile ( char *SourceFile, char *TemplateFile, char *CDFile )
                                         char            _StrLen[20 + 1];
 
                                         sprintf ( _StrLen, "%020lx",
-                                                  ( unsigned
-                                                    long )( ( CalcRealLength
+                                                  (uint32)( ( CalcRealLength
                                                               ( cs->CD_Str ) +
                                                               1 ) &
                                                             0xfffffe ) );
@@ -808,11 +807,10 @@ void CreateSourceFile ( char *SourceFile, char *TemplateFile, char *CDFile )
                                 if ( cs )
 
                                 {
-                                    char           *idstr;
-                                    long   len = 0;
+                                    char     *idstr;
+                                    uint32   len = 0;
 
                                     if ( cs->LenBytes )
-
                                     {
                                         idstr = cs->CD_Str;
                                         while ( *idstr )
@@ -830,8 +828,9 @@ void CreateSourceFile ( char *SourceFile, char *TemplateFile, char *CDFile )
                                         }
 
                                     }
+
                                     WriteString ( fpout, cs->CD_Str,
-                                                  cs->LenBytes ? len : -1,
+                                                  cs->LenBytes ? len : (uint32)-1,
                                                   cs->LenBytes );
 
                                 }
